Refactoring: concentrate data related code
