+++ src/freeserf.cc
@@ -131,122 +131,6 @@ class game_event_handler_t : public event_handler_t {
   }
 };
 
-#define MAX_DATA_PATH      1024
-
-/* Load data file from path is non-NULL, otherwise search in
-   various standard paths. */
-static int
-load_data_file(const char *path) {
-  const char *default_data_file[] = {
-    "SPAE.PA", /* English */
-    "SPAF.PA", /* French */
-    "SPAD.PA", /* German */
-    "SPAU.PA", /* Engish (US) */
-    NULL
-  };
-
-  /* Use specified path. If something was specified
-     but not found, this function should fail without
-     looking anywhere else. */
-  if (path != NULL) {
-    LOGI("main", "Looking for game data in `%s'...", path);
-    int r = data_load(path);
-    if (r < 0) return -1;
-    return 0;
-  }
-
-  /* If a path is not specified (path is NULL) then
-     the configuration file is searched for in the directories
-     specified by the XDG Base Directory Specification
-     <http://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html>.
-
-     On windows platforms the %localappdata% is used in place of XDG_CONFIG_HOME.
-  */
-
-  char cp[MAX_DATA_PATH];
-  char *env;
-
-  /* Look in home */
-  if ((env = getenv("XDG_DATA_HOME")) != NULL &&
-      env[0] != '\0') {
-    for (const char **df = default_data_file; *df != NULL; df++) {
-      snprintf(cp, sizeof(cp), "%s/freeserf/%s", env, *df);
-      LOGI("main", "Looking for game data in `%s'...", cp);
-      int r = data_load(cp);
-      if (r >= 0) return 0;
-    }
-  } else if ((env = getenv("HOME")) != NULL && env[0] != '\0') {
-    for (const char **df = default_data_file; *df != NULL; df++) {
-      snprintf(cp, sizeof(cp),
-               "%s/.local/share/freeserf/%s", env, *df);
-      LOGI("main", "Looking for game data in `%s'...", cp);
-      int r = data_load(cp);
-      if (r >= 0) return 0;
-    }
-  }
-
-#ifdef _WIN32
-  if ((env = getenv("userprofile")) != NULL && env[0] != '\0') {
-    for (const char **df = default_data_file; *df != NULL; df++) {
-      snprintf(cp, sizeof(cp),
-               "%s/.local/share/freeserf/%s", env, *df);
-      LOGI("main", "Looking for game data in `%s'...", cp);
-      int r = data_load(cp);
-      if (r >= 0) return 0;
-    }
-  }
-#endif
-
-  if ((env = getenv("XDG_DATA_DIRS")) != NULL && env[0] != '\0') {
-    char *begin = env;
-    while (1) {
-      char *end = strchr(begin, ':');
-      if (end == NULL) end = strchr(begin, '\0');
-
-      int len = static_cast<int>(end - begin);
-      if (len > 0) {
-        for (const char **df = default_data_file; *df != NULL; df++) {
-          snprintf(cp, sizeof(cp),
-                   "%.*s/freeserf/%s", len, begin, *df);
-          LOGI("main", "Looking for game data in `%s'...", cp);
-          int r = data_load(cp);
-          if (r >= 0) return 0;
-        }
-      }
-
-      if (end[0] == '\0') break;
-      begin = end + 1;
-    }
-  } else {
-    /* Look in /usr/local/share and /usr/share per XDG spec. */
-    for (const char **df = default_data_file; *df != NULL; df++) {
-      snprintf(cp, sizeof(cp),
-               "/usr/local/share/freeserf/%s", *df);
-      LOGI("main", "Looking for game data in `%s'...", cp);
-      int r = data_load(cp);
-      if (r >= 0) return 0;
-    }
-
-    for (const char **df = default_data_file; *df != NULL; df++) {
-      snprintf(cp, sizeof(cp),
-               "/usr/share/freeserf/%s", *df);
-      LOGI("main", "Looking for game data in `%s'...", cp);
-      int r = data_load(cp);
-      if (r >= 0) return 0;
-    }
-  }
-
-  /* Look in current directory */
-  for (const char **df = default_data_file; *df != NULL; df++) {
-    LOGI("main", "Looking for game data in `%s'...", *df);
-    int r = data_load(*df);
-    if (r >= 0) return 0;
-  }
-
-  return -1;
-}
-
-
 #define USAGE                                               \
   "Usage: %s [-g DATA-FILE]\n"
 #define HELP                                                \
@@ -263,8 +147,6 @@ load_data_file(const char *path) {
 
 int
 main(int argc, char *argv[]) {
-  int r;
-
   std::string data_file;
   std::string save_file;
 
@@ -337,8 +219,7 @@ main(int argc, char *argv[]) {
 
   LOGI("main", "freeserf %s", FREESERF_VERSION);
 
-  r = load_data_file(data_file.empty() ? NULL : data_file.c_str());
-  if (r < 0) {
+  if (!load_data_file(data_file)) {
     LOGE("main", "Could not load game data.");
     exit(EXIT_FAILURE);
   }
